requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
torch
Pillow
ultralytics
transformers
paddlepaddle
paddleocr
